Move Analytics to OpenTelemetryTags from AzureServiceBusTags
